python-worker/requirements.txt
# requirements.txt

# --- For the FastAPI Web Server ---
fastapi
uvicorn[standard]
python-multipart

# --- Core ML & Image Processing Libraries ---
# For Apple Silicon (M1/M2/M3), pip usually handles this correctly.
torch
torchvision
Pillow
numpy
scipy

# --- Hugging Face & Model-Specific Dependencies ---
transformers

# Dependencies specifically required by the Bria AI model's custom code
timm
kornia
